pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "openai>=2.8.1",
    "pdfplumber>=0.11.8",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-docx>=1.2.0",
    "sqlalchemy>=2.0.44",
    "werkzeug>=3.1.4",
]
